# aas_transformer.clients/aas_transformer.clients.management/src/main/resources/swagger.yml
{"openapi":"3.0.1","info":{"title":"aas_transformer.service.management","description":"REST API of AAS Transformer Management","contact":{"name":"Fraunhofer IPA","url":"https://www.ipa.fraunhofer.de","email":"benjamin.goetz@ipa.fraunhofer.de"},"version":"2.0.0-SNAPSHOT"},"servers":[{"url":"http://host.docker.internal:4010","description":"Generated server url"}],"paths":{"/metrics/processed-submodel-change-count/add-one":{"post":{"tags":["metrics-rest-controller"],"summary":"Increase processed submodel change count by one","operationId":"addOneToProcessedSubmodelChangeCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number","format":"double"}}}}}}},"/metrics/logs":{"get":{"tags":["metrics-rest-controller"],"summary":"Get transformation logs","operationId":"getTransformationLogs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformationLog"}}}}}}},"post":{"tags":["metrics-rest-controller"],"summary":"Add a transformation log","operationId":"addTransformationLog","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransformationLog"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransformationLog"}}}}}},"delete":{"tags":["metrics-rest-controller"],"summary":"Clear transformation log","operationId":"clearTransformationLog","responses":{"200":{"description":"OK"}}}},"/aas/transformer":{"get":{"tags":["transformer-rest-controller"],"summary":"Get all Transformer","operationId":"getAllTransformer","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transformer"}}}}}}},"post":{"tags":["transformer-rest-controller"],"summary":"Create Transformer","operationId":"createTransformer","parameters":[{"name":"execute","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transformer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transformer"}}}}}}},"/aas/transformer/{transformerId}/orphaned-submodels":{"post":{"tags":["transformer-rest-controller"],"summary":"Get orphaned destination submodel IDs of a Transformer. Might return empty list even if orphans exist.","operationId":"getOrphanedDestinationSubmodels","parameters":[{"name":"transformerId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Submodel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/transformer-jobs/waiting":{"get":{"tags":["job-rest-controller"],"summary":"Get all waiting jobs","operationId":"getWaitingJobs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformationJob"}}}}}}}},"/transformer-jobs/in-progress":{"get":{"tags":["job-rest-controller"],"summary":"Get all in-progress jobs","operationId":"getInProgressJobs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformationJob"}}}}}}}},"/transformer-jobs/count/waiting":{"get":{"tags":["job-rest-controller"],"summary":"Get count of waiting jobs","operationId":"getWaitingJobCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/transformer-jobs/count/total":{"get":{"tags":["job-rest-controller"],"summary":"Get count of all jobs","operationId":"getTotalJobCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/transformer-jobs/count/in-progress":{"get":{"tags":["job-rest-controller"],"summary":"Get count of in-progress jobs","operationId":"getInProgressJobCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/metrics/submodel-change-count":{"get":{"tags":["metrics-rest-controller"],"summary":"Get the number of submodel changes","operationId":"getSubmodelChangeCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number","format":"double"}}}}}}},"/metrics/processed-submodel-change-count":{"get":{"tags":["metrics-rest-controller"],"summary":"Get the number of processed submodel changes","operationId":"getProcessedSubmodelChangeCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number","format":"double"}}}}}}},"/aas/transformer/{transformerId}":{"get":{"tags":["transformer-rest-controller"],"summary":"Get one Transformer","operationId":"getTransformer","parameters":[{"name":"transformerId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transformer"}}}}}},"delete":{"tags":["transformer-rest-controller"],"summary":"Delete one Transformer","operationId":"deleteTransformer","parameters":[{"name":"transformerId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"doCleanup","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK"}}}},"/aas/transformer/{transformerId}/destination-submodels":{"get":{"tags":["transformer-rest-controller"],"summary":"Get destination submodel IDs of a Transformer (destination definition must have no templates)","operationId":"getDestinationSubmodels","parameters":[{"name":"transformerId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/aas/transformer/events":{"get":{"tags":["transformer-rest-controller"],"summary":"Get Stream of Transformer Change Events","operationId":"getTransformerChangeEventStream","responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformerChangeEvent"}}}}}}}},"/aas/transformer/events/dto-listener":{"get":{"tags":["transformer-rest-controller"],"summary":"Get Stream of Transformer Change Events as DTOListener","operationId":"getTransformerChangeEventDTOListenerStream","responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformerChangeEventDTOListener"}}}}}}}},"/aas/transformer/dto-listener":{"get":{"tags":["transformer-rest-controller"],"summary":"Get all Transformer as DTOListener","operationId":"getAllTransformerDTOListener","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformerDTOListener"}}}}}}}}},"components":{"schemas":{"TransformationLog":{"required":["date","destinationAasId","destinationSubmodelId","sourceSubmodelId"],"type":"object","properties":{"destinationAasId":{"type":"string"},"destinationSubmodelId":{"type":"string"},"sourceSubmodelId":{"type":"string"},"executorId":{"type":"string","format":"uuid"},"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date-time"}}},"Destination":{"type":"object","properties":{"aasDestination":{"$ref":"#/components/schemas/DestinationAAS"},"submodelDestination":{"$ref":"#/components/schemas/DestinationSubmodel"}}},"DestinationAAS":{"type":"object","properties":{"id":{"type":"string"}}},"DestinationSubmodel":{"required":["id","idShort"],"type":"object","properties":{"idShort":{"type":"string"},"id":{"type":"string"}}},"RuleOperator":{"type":"string","enum":["EQUALS","NOT_EQUALS"]},"SourceSubmodelIdRule":{"type":"object","properties":{"ruleOperator":{"$ref":"#/components/schemas/RuleOperator"},"sourceSubmodelId":{"$ref":"#/components/schemas/SubmodelId"}}},"SubmodelId":{"required":["id"],"type":"object","properties":{"type":{"$ref":"#/components/schemas/SubmodelIdType"},"id":{"type":"string"}}},"SubmodelIdType":{"type":"string","enum":["ID","ID_SHORT","SEMANTIC_ID"]},"Transformer":{"required":["destination","id","sourceSubmodelIdRules","transformerActions","version"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"destination":{"$ref":"#/components/schemas/Destination"},"transformerActions":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/TransformerActionCopy"},{"$ref":"#/components/schemas/TransformerActionSubmodelElementTemplate"},{"$ref":"#/components/schemas/TransformerActionSubmodelTemplate"},{"$ref":"#/components/schemas/TransformerActionTsAvg"},{"$ref":"#/components/schemas/TransformerActionTsMdn"}]}},"sourceSubmodelIdRules":{"type":"array","items":{"$ref":"#/components/schemas/SourceSubmodelIdRule"}},"version":{"type":"integer","format":"int32"}}},"TransformerAction":{"type":"object","properties":{"actionType":{"$ref":"#/components/schemas/TransformerActionType"}},"discriminator":{"propertyName":"actionType"}},"TransformerActionCopy":{"required":["sourceSubmodelElement"],"type":"object","allOf":[{"$ref":"#/components/schemas/TransformerAction"},{"type":"object","properties":{"sourceSubmodelElement":{"type":"string"},"destinationSubmodelElement":{"type":"string"}}}]},"TransformerActionSubmodelElementTemplate":{"required":["destinationSubmodelElementId","submodelElementTemplate"],"type":"object","allOf":[{"$ref":"#/components/schemas/TransformerAction"},{"type":"object","properties":{"destinationSubmodelElementId":{"type":"string"},"submodelElementTemplate":{"type":"string"}}}]},"TransformerActionSubmodelTemplate":{"required":["submodelTemplate"],"type":"object","allOf":[{"$ref":"#/components/schemas/TransformerAction"},{"type":"object","properties":{"submodelTemplate":{"type":"string"}}}]},"TransformerActionTsAvg":{"required":["filterSize","recordNames","transformerActionType"],"type":"object","allOf":[{"$ref":"#/components/schemas/TransformerAction"},{"type":"object","properties":{"recordNames":{"type":"array","items":{"type":"string"}},"filterSize":{"type":"integer","format":"int32"},"transformerActionType":{"$ref":"#/components/schemas/TransformerActionType"}}}]},"TransformerActionTsMdn":{"required":["filterSize","recordNames","transformerActionType"],"type":"object","allOf":[{"$ref":"#/components/schemas/TransformerAction"},{"type":"object","properties":{"recordNames":{"type":"array","items":{"type":"string"}},"filterSize":{"type":"integer","format":"int32"},"transformerActionType":{"$ref":"#/components/schemas/TransformerActionType"}}}]},"TransformerActionType":{"type":"string","enum":["COPY","SUBMODEL_TEMPLATE","SUBMODEL_ELEMENT_TEMPLATE","TS_AVG","TS_MDN","TS_TAKE_EVERY","TS_DROP_EVERY"]},"AdministrativeInformation":{"type":"object","properties":{"version":{"type":"string"},"revision":{"type":"string"},"templateId":{"type":"string"},"creator":{"$ref":"#/components/schemas/Reference"},"embeddedDataSpecifications":{"type":"array","items":{"$ref":"#/components/schemas/EmbeddedDataSpecification"}}}},"DataSpecificationContent":{"type":"object"},"DataTypeDefXsd":{"type":"string","enum":["ANY_URI","BASE64BINARY","BOOLEAN","BYTE","DATE","DATE_TIME","DECIMAL","DOUBLE","DURATION","FLOAT","GDAY","GMONTH","GMONTH_DAY","GYEAR","GYEAR_MONTH","HEX_BINARY","INT","INTEGER","LONG","NEGATIVE_INTEGER","NON_NEGATIVE_INTEGER","NON_POSITIVE_INTEGER","POSITIVE_INTEGER","SHORT","STRING","TIME","UNSIGNED_BYTE","UNSIGNED_INT","UNSIGNED_LONG","UNSIGNED_SHORT"]},"EmbeddedDataSpecification":{"type":"object","properties":{"dataSpecificationContent":{"$ref":"#/components/schemas/DataSpecificationContent"},"dataSpecification":{"$ref":"#/components/schemas/Reference"}}},"Extension":{"type":"object","properties":{"name":{"type":"string"},"value":{"type":"string"},"valueType":{"$ref":"#/components/schemas/DataTypeDefXsd"},"refersTo":{"type":"array","items":{"$ref":"#/components/schemas/Reference"}},"supplementalSemanticIds":{"type":"array","items":{"$ref":"#/components/schemas/Reference"}},"semanticId":{"$ref":"#/components/schemas/Reference"}}},"Key":{"type":"object","properties":{"value":{"type":"string"},"type":{"$ref":"#/components/schemas/KeyTypes"}}},"KeyTypes":{"type":"string","enum":["ANNOTATED_RELATIONSHIP_ELEMENT","ASSET_ADMINISTRATION_SHELL","BASIC_EVENT_ELEMENT","BLOB","CAPABILITY","CONCEPT_DESCRIPTION","DATA_ELEMENT","ENTITY","EVENT_ELEMENT","FILE","FRAGMENT_REFERENCE","GLOBAL_REFERENCE","IDENTIFIABLE","MULTI_LANGUAGE_PROPERTY","OPERATION","PROPERTY","RANGE","REFERABLE","REFERENCE_ELEMENT","RELATIONSHIP_ELEMENT","SUBMODEL","SUBMODEL_ELEMENT","SUBMODEL_ELEMENT_COLLECTION","SUBMODEL_ELEMENT_LIST"]},"LangStringNameType":{"type":"object","properties":{"language":{"type":"string"},"text":{"type":"string"}}},"LangStringTextType":{"type":"object","properties":{"language":{"type":"string"},"text":{"type":"string"}}},"ModellingKind":{"type":"string","enum":["INSTANCE","TEMPLATE"]},"Qualifier":{"type":"object","properties":{"value":{"type":"string"},"type":{"type":"string"},"kind":{"$ref":"#/components/schemas/QualifierKind"},"valueType":{"$ref":"#/components/schemas/DataTypeDefXsd"},"valueId":{"$ref":"#/components/schemas/Reference"},"supplementalSemanticIds":{"type":"array","items":{"$ref":"#/components/schemas/Reference"}},"semanticId":{"$ref":"#/components/schemas/Reference"}}},"QualifierKind":{"type":"string","enum":["CONCEPT_QUALIFIER","TEMPLATE_QUALIFIER","VALUE_QUALIFIER"]},"Reference":{"type":"object","properties":{"type":{"$ref":"#/components/schemas/ReferenceTypes"},"keys":{"type":"array","items":{"$ref":"#/components/schemas/Key"}},"referredSemanticId":{"$ref":"#/components/schemas/Reference"}}},"ReferenceTypes":{"type":"string","enum":["EXTERNAL_REFERENCE","MODEL_REFERENCE"]},"Submodel":{"type":"object","properties":{"submodelElements":{"type":"array","items":{"$ref":"#/components/schemas/SubmodelElement"}},"id":{"type":"string"},"administration":{"$ref":"#/components/schemas/AdministrativeInformation"},"displayName":{"type":"array","items":{"$ref":"#/components/schemas/LangStringNameType"}},"description":{"type":"array","items":{"$ref":"#/components/schemas/LangStringTextType"}},"idShort":{"type":"string"},"category":{"type":"string"},"extensions":{"type":"array","items":{"$ref":"#/components/schemas/Extension"}},"embeddedDataSpecifications":{"type":"array","items":{"$ref":"#/components/schemas/EmbeddedDataSpecification"}},"supplementalSemanticIds":{"type":"array","items":{"$ref":"#/components/schemas/Reference"}},"semanticId":{"$ref":"#/components/schemas/Reference"},"qualifiers":{"type":"array","items":{"$ref":"#/components/schemas/Qualifier"}},"kind":{"$ref":"#/components/schemas/ModellingKind"}}},"SubmodelElement":{"type":"object","properties":{"embeddedDataSpecifications":{"type":"array","items":{"$ref":"#/components/schemas/EmbeddedDataSpecification"}},"displayName":{"type":"array","items":{"$ref":"#/components/schemas/LangStringNameType"}},"description":{"type":"array","items":{"$ref":"#/components/schemas/LangStringTextType"}},"idShort":{"type":"string"},"category":{"type":"string"},"extensions":{"type":"array","items":{"$ref":"#/components/schemas/Extension"}},"supplementalSemanticIds":{"type":"array","items":{"$ref":"#/components/schemas/Reference"}},"semanticId":{"$ref":"#/components/schemas/Reference"},"qualifiers":{"type":"array","items":{"$ref":"#/components/schemas/Qualifier"}}}},"TransformationJob":{"required":["transformationJobAction"],"type":"object","properties":{"transformationJobAction":{"$ref":"#/components/schemas/TransformationJobAction"},"transformerId":{"type":"string","format":"uuid"},"submodelId":{"type":"string"},"submodel":{"$ref":"#/components/schemas/Submodel"}}},"TransformationJobAction":{"type":"string","enum":["EXECUTE","DELETE"]},"TransformerChangeEvent":{"type":"object","properties":{"type":{"$ref":"#/components/schemas/TransformerChangeEventType"},"transformer":{"$ref":"#/components/schemas/Transformer"}}},"TransformerChangeEventType":{"type":"string","enum":["CREATE","UPDATE","DELETE"]},"TransformerChangeEventDTOListener":{"type":"object","properties":{"type":{"$ref":"#/components/schemas/TransformerChangeEventType"},"transformer":{"$ref":"#/components/schemas/TransformerDTOListener"}}},"TransformerDTOListener":{"required":["sourceSubmodelIdRules"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"sourceSubmodelIdRules":{"type":"array","items":{"$ref":"#/components/schemas/SourceSubmodelIdRule"}}}}}}}